Revert "Change build system to use CMake. This enables out-of-tree compilatio…"
