+++ module/Kbuild
@@ -1,4 +1,13 @@
-ccflags-y := -I$(src)/include
+NVIDIA_DRIVER_VER ?= $(shell nvidia-smi --query-gpu=driver_version --format=csv,noheader | head -1)
 
-obj-m := bafs-core.o
-bafs-core-y := bafs/core.o
+NVIDIA_DRIVER_DIR ?= /usr/src/nvidia-$(NVIDIA_DRIVER_VER)
+NVIDIA_INCLUDE_DIR ?= /usr/src/nvidia-$(NVIDIA_DRIVER_VER)/nvidia
+
+BAFS_INCLUDE_DIR ?= $(src)/include
+
+obj-m := bafs_core.o
+
+ccflags-y += -I$(NVIDIA_INCLUDE_DIR)
+ccflags-y += -I$(BAFS_INCLUDE_DIR)
+
+KBUILD_EXTRA_SYMBOLS := $(NVIDIA_DRIVER_DIR)/Module.symvers
