+++ module/include/bafs_common.h
@@ -1,0 +1,14 @@
+#ifndef __BAFS_COMMON_H__
+#define __BAFS_COMMON_H__
+
+typedef uint32_t      bafs_mem_hnd_t;
+typedef uint32_t      bafs_group_hnd_t;
+typedef unsigned long addr_;
+
+enum LOC {
+    CPU,
+    CUDA
+};
+
+
+#endif                          // __BAFS_COMMON_H__
