+++ module/include/bafs_types.h
@@ -1,5 +1,5 @@
-#ifndef _LINUX_BAFS_TYPES_H_
-#define _LINUX_BAFS_TYPES_H_
+#ifndef __BAFS_TYPES_H__
+#define __BAFS_TYPES_H__
 
 
 #include <linux/mm.h>
@@ -9,7 +9,7 @@
 #include <linux/dma-mapping.h>
 #include <linux/pci.h>
 
-#include <linux/bafs.h>
+#include "bafs_common.h"
 
 #define PCI_CLASS_NVME      0x010802
 #define PCI_CLASS_NVME_MASK 0xffffff
@@ -108,7 +108,7 @@ struct bafs_mem {
     struct list_head         dma_list;
     struct kref              ref;
     bafs_mem_hnd_t           mem_id;
-    unsigned                 loc;
+    enum LOC                 loc;
     enum STATE               state;
     unsigned long            vaddr;
     unsigned long            size;
@@ -130,7 +130,7 @@ struct bafs_mem_dma {
     struct bafs_ctrl*         ctrl;
     nvidia_p2p_dma_mapping_t* cuda_mapping;
     unsigned long             n_addrs;
-    unsigned long *           addrs;
+    addr_*               addrs;
     unsigned                  map_gran;
 
 };
